requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
Markdown==3.5.2
google-cloud-aiplatform==1.38.1
Flask-WTF
google-genai




